tests: the wide-constant case holds its constant in long long on every target
constant_set_again_after_call.c keeps a 64-bit constant in `long` in
its wide-constant case, but `long` is 32 bits on the LLP64 Windows
targets: `long k = 0x123456789abLL` truncates to 0x456789ab and the
case returns a 32-bit result, so `main`'s comparison against the
64-bit expression always fails and the program exits 5. clang
targeting x86_64-pc-windows-msvc reports both, the conversion
(-Wconstant-conversion, "changes value from 1250999896491 to
1164413355") and the comparison (-Wtautological-constant-out-of-range-
compare, "always true"). NATIVE_PE_X64_FIXTURES failed on the
windows-x64 box at -O0 and -O with exit 5; the code badc generates
matches what clang's reading of the source requires.

The constant and the case's result are `long long`, which is the
same 64-bit type as `long` on the LP64 targets the snapshots cover.
Built by badc for windows-x64 and windows-aarch64 at -O0 and -O, the
fixture exits 0 on both Windows boxes.

+++ tests/fixtures/c/constant_set_again_after_call.c
@@ -47,8 +47,8 @@ __attribute__((noinline)) static long returned_constant(long x) {
     return r;
 }
 
-__attribute__((noinline)) static long wide_constant(long x) {
-    long k = 0x123456789abLL;
+__attribute__((noinline)) static long long wide_constant(long x) {
+    long long k = 0x123456789abLL;
     long a = note(x);
     long b = note(a);
     return (k ^ b) + (k >> 4);
